A Very Big Sum

+++ A Very Big Sum/src/Solution.java
@@ -5,10 +5,9 @@ public class Solution {
     public static void main(String[] args) {
         Scanner in = new Scanner(System.in);
         int n = in.nextInt();
-        int result = 0;
+        long result = 0;
         for (int i = 0; i < n; i++) {
             result += in.nextInt();
-
         }
         System.out.println(result);
     }
